% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{add_form_event_structure}
\alias{add_form_event_structure}
\title{Add event-form level repeating structure to event mappings}
\usage{
add_form_event_structure(data, db_instrument_repeating)
}
\arguments{
\item{data}{a tibble containing event-instrument mappings.}

\item{db_instrument_repeating}{a tibble of repeating instrument settings
created by \code{\link[=pull_instrument_repeating]{pull_instrument_repeating()}}.}
}
\value{
The original event-instrument mapping tibble supplemented with
\code{custom_form_label} and \code{repeat_structure} columns.
}
\description{
Add event-form level repeating structure to event mappings
}
\details{
This function merges repeating instrument information from REDCap with
event-instrument mappings. Repeating together vs. separate is inferred
from whether \code{form} is populated.
}
\keyword{internal}
