% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{add_metadata_sheet}
\alias{add_metadata_sheet}
\title{Add the metadata sheet}
\usage{
add_metadata_sheet(
  supertbl,
  supertbl_meta,
  wb,
  add_labelled_column_headers,
  table_style,
  column_width,
  na_replace
)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{wb}{An \code{openxlsx2} workbook object}

\item{add_labelled_column_headers}{Whether or not to include labelled outputs.}

\item{table_style}{Any excel table style name or "none" (see "formatting"
vignette in \link[openxlsx2]{wb_add_data_table}). Default "tableStyleLight8".}

\item{column_width}{Width to set columns across the workbook. Default
"auto", otherwise a numeric value. Standard Excel is 8.43.}

\item{na_replace}{The value used to replace \code{NA} values in \code{supertbl}. The default is "".}

\item{superbl_meta}{an \code{unnest}-ed metadata tibble from the supertibble}
}
\value{
A dataframe
}
\description{
Internal helper function. Adds appropriate elements to \code{wb} object. Returns
a dataframe.
}
\keyword{internal}
