% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_partial_keys}
\alias{add_partial_keys}
\title{Add Partial Keys for Helper Variables}
\usage{
add_partial_keys(db_data)
}
\arguments{
\item{db_data}{A REDCap database export}
}
\description{
Make helper vars \code{redcap_event} and \code{redcap_arm} available as offshoots from \code{redcap_event_name}.
}
\keyword{internal}
