% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{add_skimr_metadata}
\alias{add_skimr_metadata}
\title{Add \code{skimr} metrics to supertibble metadata}
\usage{
add_skimr_metadata(supertbl)
}
\arguments{
\item{supertbl}{a supertibble object}
}
\value{
A supertibble object
}
\description{
Utility function to support adding \code{skimr} metrics to supertibble metadata
rows.
}
\keyword{internal}
