% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_labs_factor}
\alias{apply_labs_factor}
\title{Apply factor labels to a vector}
\usage{
apply_labs_factor(x, labels, ...)
}
\arguments{
\item{x}{a vector to label}

\item{labels}{a named vector of labels in the format \code{c(value = label)}}

\item{\dots}{unused, needed to ignore extra arguments that may be passed}
}
\value{
factor
}
\description{
Apply factor labels to a vector
}
\details{
Dots are needed to ignore \code{ptype} argument that may be passed to \code{apply_labs_haven}
}
\keyword{internal}
