% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{as_supertbl}
\alias{as_supertbl}
\title{Add supertbl S3 class}
\usage{
as_supertbl(x)
}
\arguments{
\item{x}{an object to class}
}
\value{
The object with \code{redcaptidier_supertbl} S3 class
}
\description{
Add supertbl S3 class
}
\keyword{internal}
