% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{bind_supertbl_metadata}
\alias{bind_supertbl_metadata}
\title{Bind supertbl metadata}
\usage{
bind_supertbl_metadata(supertbl)
}
\arguments{
\item{supertbl}{A supertibble generated using \code{read_redcap()}}
}
\description{
Simple helper function for binding supertbl metadata into one table. This
supports creating the metadata XLSX sheet as well as \code{supertbl_recode}.
}
\keyword{internal}
