% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tables.R
\name{bind_tables}
\alias{bind_tables}
\title{Bind REDCap Tidy Tibbles to Specified Environments}
\usage{
bind_tables(
  .data,
  environment = global_env(),
  redcap_form_name = NULL,
  structure = NULL
)
}
\arguments{
\item{.data}{A tidy table provided by \code{read_redcap_tidy()}}

\item{environment}{The environment to assign the tidy data (default \code{global_env()}). For new environment, it is recommended to use \code{rlang::new_environment()}.}

\item{redcap_form_name}{Specify REDCap form names to load into environment. Default behavior is all forms.}

\item{structure}{Specify REDCap structure to load into environment, either \code{repeating} or \code{nonrepeating}. Defaults to all structures.}
}
\description{
In Development
}
