% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_tibbles.R
\name{build_by}
\alias{build_by}
\title{Intelligently retrieve the join by cols}
\usage{
build_by(supertbl, x, y, is_mixed)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{read_redcap()}. Required.}

\item{x, y}{A pair of data frames, data frame extensions (e.g. a tibble), or
lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{is_mixed}{TRUE/FALSE, whether or not the given tables contain a mixed structure}
}
\value{
a character vector
}
\description{
Intelligently retrieve the join by cols
}
\keyword{interal}
