% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_arg_is_logical}
\alias{check_arg_is_logical}
\title{Check that an argument is logical}
\usage{
check_arg_is_logical(x, ..., arg = caller_arg(x))
}
\arguments{
\item{x}{An object to check}

\item{...}{additional arguments passed on to checkmate}

\item{arg}{The name of the argument to include in an error message. Captured
by \code{rlang::caller_arg()} by default}
}
\value{
\code{TRUE} if \code{x} is a logical vector. An error message otherwise
}
\description{
Check that an argument is logical
}
\keyword{internal}
