% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_equal_col_summaries}
\alias{check_equal_col_summaries}
\title{Check equal distinct values between two columns}
\usage{
check_equal_col_summaries(data, col1, col2, call = caller_env())
}
\arguments{
\item{data}{a dataframe}

\item{col1}{a column to group by}

\item{col2}{a column to check for uniqueness}
}
\description{
Takes a dataframe and two columns and checks if \code{\link[=n_distinct]{n_distinct()}} of the second
column is all unique based on grouping of the first column.
}
\keyword{internal}
