% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_extra_field_values}
\alias{check_extra_field_values}
\title{Check data field for field values not in metadata}
\usage{
check_extra_field_values(x, values)
}
\arguments{
\item{x}{data field}

\item{values}{expected field values}
}
\description{
Check data field for field values not in metadata
}
\keyword{internal}
