% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_field_is_logical}
\alias{check_field_is_logical}
\title{Parse logical field and warn if parsing errors occurred}
\usage{
check_field_is_logical(x, field_name, field_type, call = caller_env())
}
\arguments{
\item{x}{vector to parse}

\item{field_name}{field name for warning message}

\item{field_type}{field type for warning message}

\item{call}{call for warning message}
}
\description{
Parse logical field and warn if parsing errors occurred
}
\keyword{internal}
