% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fields_are_checkboxes}
\alias{check_fields_are_checkboxes}
\title{Check fields are of checkbox field type}
\usage{
check_fields_are_checkboxes(metadata_tbl, call = caller_env())
}
\arguments{
\item{metadata_tbl}{A metadata tibble from a supertibble}

\item{call}{The calling environment to use in the error message}
}
\description{
Check fields are of checkbox field type
}
\keyword{internal}
