% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fields_exist}
\alias{check_fields_exist}
\title{Check fields exist for checkbox combination}
\usage{
check_fields_exist(fields, expr, call = caller_env())
}
\arguments{
\item{fields}{Vector of character strings to check the length of}

\item{expr}{An expression}

\item{call}{The calling environment to use in the error message}
}
\description{
Check fields exist for checkbox combination
}
\keyword{internal}
