% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check if file already exists}
\usage{
check_file_exists(file, overwrite, call = caller_env())
}
\arguments{
\item{file}{The file that is being checked}

\item{overwrite}{Whether the file was declared to be overwritten}

\item{call}{The calling environment to use in the error message}
}
\value{
An error message saying the requested file already exists
}
\description{
Provide an error message when a file is declared for writing that already
exists.
}
\details{
In the case of \code{write_redcap_xlsx()}, this should only error when a file
already exists and is not declared for \code{overwite}.
}
\keyword{internal}
