% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_forms_exist}
\alias{check_forms_exist}
\title{Check that all requested instruments are in REDCap project metadata}
\usage{
check_forms_exist(db_metadata, forms, call = caller_env())
}
\arguments{
\item{db_metadata}{The metadata file read by
\code{REDCapR::redcap_metadata_read()}}

\item{forms}{The character vector of instrument names passed to
\code{read_redcap()}}

\item{call}{the calling environment to use in the error message}
}
\value{
An error message listing the requested instruments that don't exist
}
\description{
Provide an error message when any instrument names are passed to
\code{read_redcap()} that do not exist in the project metadata.
}
\keyword{internal}
