% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{check_labelled}
\alias{check_labelled}
\title{Check if labelled}
\usage{
check_labelled(supertbl, add_labelled_column_headers, call = caller_env())
}
\arguments{
\item{supertbl}{a supertibble generated using \code{\link[=read_redcap]{read_redcap()}}}

\item{add_labelled_column_headers}{Whether or not to include labelled outputs}

\item{call}{the calling environment to use in the warning message}
}
\value{
A boolean
}
\description{
Checks if a supplied supertibble is labelled and throws an error if it is not
but \code{labelled} is set to \code{TRUE}
}
\keyword{internal}
