% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_metadata_field_types}
\alias{check_metadata_field_types}
\title{Check metadata field types against parsed data types}
\usage{
check_metadata_field_types(db_data, db_metadata, call = caller_env())
}
\arguments{
\item{db_data}{A REDCap database object}

\item{db_metadata}{A REDCap metadata object}

\item{call}{The calling environment to use in the warning message}
}
\description{
Check metadata field types against parsed data types
}
\keyword{internal}
