% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_metadata_fields_exist}
\alias{check_metadata_fields_exist}
\title{Check metadata fields exist for checkbox combination}
\usage{
check_metadata_fields_exist(metadata_tbl, cols, call = caller_env())
}
\arguments{
\item{metadata_tbl}{A metadata tibble from the supertibble generated by \code{\link[=read_redcap]{read_redcap()}}.}

\item{cols}{Selected columns identified for \code{\link[=combine_checkboxes]{combine_checkboxes()}} to be
cross checked against \code{metadata_tibble$field_name}}

\item{call}{The calling environment to use in the error message}
}
\description{
Similar to \code{\link[=check_fields_exist]{check_fields_exist()}}, but instead of verifying fields that exist
in the data tibble this seeks to verify their existence under the metadata
tibble \code{field_name}s.
}
\keyword{internal}
