% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_parsed_labels}
\alias{check_parsed_labels}
\title{Check that parsed labels are not duplicated}
\usage{
check_parsed_labels(parsed_labs, field_name, call = caller_env(n = 2))
}
\arguments{
\item{parsed_labs}{a vector of parsed labels produced by \code{parse_labels()}}

\item{field_name}{the name of the field associated with the labels to use in the warning message}

\item{call}{the calling environment to use in the error message. The parent of calling environment
by default because this check usually occurs 2 frames below the relevant context for the user}
}
\value{
a warning message alerting specifying the duplicate labels and REDCap field affected
}
\description{
Check that parsed labels are not duplicated
}
\keyword{internal}
