% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_redcap_populated}
\alias{check_redcap_populated}
\title{Check that REDCap Database is Populated}
\usage{
check_redcap_populated(db_data)
}
\arguments{
\item{db_data}{The REDCap database output generated by \code{REDCapR::redcap_read_oneshot()$data}}
}
\description{
Check for potential outputs where metadata is present but \code{nrow} and
\code{ncol} equal `0`. This will cause \code{multi_choice_to_labels} to fail,
but a helpeful error message should be provided.
}
\keyword{internal}
