% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_repeat_and_nonrepeat}
\alias{check_repeat_and_nonrepeat}
\title{Check for instruments that have both repeating and non-repeating structure}
\usage{
check_repeat_and_nonrepeat(db_data)
}
\arguments{
\item{db_data}{The REDCap database output generated by
\code{REDCapR::redcap_read_oneshot()$data}}
}
\value{
A helpful error message alerting the user to existence of a form containing
both repeating and non-repeating structure.
}
\description{
Check for potential instruments that are given both repeating and
nonrepeating structure. \code{REDCapTidieR} does not support database
structures built this way.
}
\keyword{internal}
