% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_unmapped_repeating_forms}
\alias{check_unmapped_repeating_forms}
\title{Check for repeating forms not mapped to events}
\usage{
check_unmapped_repeating_forms(
  db_event_instruments,
  db_instrument_repeating,
  call = caller_env()
)
}
\arguments{
\item{db_event_instruments}{a tibble of event-instrument mappings created by
\code{\link[=pull_event_instruments]{pull_event_instruments()}}.}

\item{db_instrument_repeating}{a tibble of repeating instrument settings
created by \code{\link[=pull_instrument_repeating]{pull_instrument_repeating()}}.}

\item{call}{The calling environment to use in the error message}
}
\description{
Check for repeating forms not mapped to events
}
\keyword{internal}
