% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_user_rights}
\alias{check_user_rights}
\title{Check for API User Rights Errors}
\usage{
check_user_rights(db_data, db_metadata)
}
\arguments{
\item{db_data}{The REDCap database output defined by \code{REDCapR::reedcap_read_oneshot()$data}}

\item{db_metadata}{The REDCap metadata output defined by \code{REDCapR::redcap_metadata_read()$data}}
}
\description{
Check for potential user access rights errors and supply user with an appropriate warning message. this can occur when metadata forms/field names do not appear in a database export.
}
\keyword{internal}
