% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_valid_api_token}
\alias{check_valid_api_token}
\title{Check for invalid API token supplied}
\usage{
check_valid_api_token(db_metadata, call = caller_env())
}
\arguments{
\item{db_metadata}{The REDCap metadata output generated by \code{REDCapR::redcap_metadata_read()}}

\item{call}{the calling environment to use in the warning message}
}
\value{
A helpful error message alerting the user of an invalid API token.
}
\description{
Check for when a user provides a valid 32 length hexadecimal that conforms to
checkmate checks, but the API key itself is not valid according to the
database user registry.
}
\keyword{internal}
