% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_values_to_length}
\alias{check_values_to_length}
\title{Check values_to length against detected number of checkbox fields}
\usage{
check_values_to_length(col_groups, values_to, call = caller_env())
}
\arguments{
\item{col_groups}{a list of column groups identified by checkbox field detection}

\item{values_to}{a user defined character vector passed from \code{\link[=combine_checkboxes]{combine_checkboxes()}}}

\item{call}{The calling environment to use in the error message}
}
\description{
Check values_to length against detected number of checkbox fields
}
\keyword{internal}
