% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkbox_appender}
\alias{checkbox_appender}
\title{Checkbox String Appender Helper Function}
\usage{
checkbox_appender(field_name, string, raw_or_label)
}
\arguments{
\item{field_name}{The \code{db_metadata$field_name} to append onto the string}

\item{string}{A \code{db_metadata$select_choices_or_calculations} field pre-filtered for checkbox \code{field_type}}

\item{raw_or_label}{A string (either 'raw' or 'label') that specifies whether to export the raw coded values or the labels for the options of multiple choice fields.}
}
\description{
Takes a \code{db_metadata$select_choices_or_calculations} field pre-filtered for checkbox \code{field_type} and returns a vector of key names for appending to checkbox variables.
}
\keyword{internal}
