% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_multi_to_single_column.R
\name{combine_checkboxes}
\alias{combine_checkboxes}
\title{Combine Checkbox Fields into a Single Column}
\usage{
combine_checkboxes(
  supertbl,
  tbl,
  cols,
  values_to,
  multi_value_label = "Multiple",
  values_fill = NA,
  raw_or_label = "label",
  keep = TRUE
)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{\link[=read_redcap]{read_redcap()}}. Required.}

\item{tbl}{The \code{redcap_form_name} of the data tibble to extract. Required.}

\item{cols}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Checbox columns to combine to
single column. Required.}

\item{values_to}{A string specifying the name of the column to combine checkbox
values under. Required.}

\item{multi_value_label}{A string specifying the value to be used when multiple
checkbox fields are selected. Default "Multiple".}

\item{values_fill}{Value to use when no checkboxes are selected. Default \code{NA}.}

\item{raw_or_label}{Either 'raw' or 'label' to specify whether to use raw coded
values or labels for the options. Default 'label'.}

\item{keep}{Logical indicating whether to keep the original checkbox fields in
the output. Default \code{TRUE}.}
}
\value{
A modified supertibble.
}
\description{
\code{\link[=combine_checkboxes]{combine_checkboxes()}} consolidates multiple checkbox fields in a REDCap data
tibble into a single column. This transformation simplifies analysis by
combining multiple binary columns into a singular and informative labelled
factor column.
}
\examples{
\dontrun{
supertbl <- read_redcap(redcap_uri, token)
combined_tbl <- combine_checkboxes(
  supertbl = supertbl,
  tbl = "demographics",
  cols = starts_with("race"),
  values_to = "race_combined",
  multi_value_label = "Multiple",
  values_fill = NA
)
}

}
