% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_checkboxes.R
\name{convert_metadata_spec}
\alias{convert_metadata_spec}
\title{Use metadata_spec to convert new column values}
\usage{
convert_metadata_spec(
  .new_col_item,
  metadata_spec,
  data_tbl_mod,
  raw_or_label,
  multi_value_label,
  values_fill
)
}
\arguments{
\item{.new_col_item}{A character string}

\item{metadata_spec}{A tibble output from \code{\link[=convert_metadata_spec]{convert_metadata_spec()}}}

\item{data_tbl_mod}{A modified data tibble}

\item{raw_or_label}{Either 'raw' or 'label' to specify whether to use raw coded
values or labels for the options. Default 'label'.}

\item{multi_value_label}{A string specifying the value to be used when multiple
checkbox fields are selected. Default "Multiple".}

\item{values_fill}{Value to use when no checkboxes are selected. Default \code{NA}.}
}
\value{
a tibble
}
\description{
\code{\link[=convert_metadata_spec]{convert_metadata_spec()}} uses the \code{metadata_spec} table provided by \code{\link[=get_metadata_spec]{get_metadata_spec()}}
to automatically convert new column values to either:
\itemize{
\item A \code{raw_or_label} checkbox value when only a single value is detected
\item \code{mult_value_label} when multiple values are detected
\item \code{values_fill} when \code{NA} is detected
}
}
\keyword{internal}
