% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap_long.R
\name{convert_mixed_instrument}
\alias{convert_mixed_instrument}
\title{Convert Mixed Structure Instruments to Repeating Instruments}
\usage{
convert_mixed_instrument(db_data_long, mixed_structure_ref)
}
\arguments{
\item{db_data_long}{The longitudinal REDCap database output defined by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{mixed_structure_ref}{Reference dataframe containing mixed structure
fields and forms.}
}
\value{
Returns a \code{tibble} with list elements containing tidy dataframes. Users
can access dataframes under the \code{redcap_data} column with reference to
\code{form_name} and \code{structure} column details.
}
\description{
For longitudinal projects where users set \code{allow_mixed_structure} to \code{TRUE},
this function will handle the process of setting the nonrepeating parts of the
instrument to repeating ones with a single instance.
}
\keyword{internal}
