% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_repeat_instance_vars}
\alias{create_repeat_instance_vars}
\title{Utility function to convert redcap repeat instance columns into appropriate
form and event columns}
\usage{
create_repeat_instance_vars(db_data)
}
\arguments{
\item{db_data}{The REDCap database output generated by
\code{REDCapR::redcap_read_oneshot()$data}}
}
\value{
A dataframe.
}
\description{
Utility function to convert redcap repeat instance columns into appropriate
form and event columns
}
\details{
The output of a standard REDCap export with repeating forms and/or events
makes use of \code{redcap_repeat_instance} in combination with
\code{redcap_repeat_instrument} and whether or not data exists in both. Instead,
rename and separate \code{redcap_repeat_instance} into \code{redcap_form_instance} and
\code{redcap_event_instance}.
}
\keyword{internal}
