% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_has_repeat_forms}
\alias{db_has_repeat_forms}
\title{Check whether a REDCap database has repeat forms}
\usage{
db_has_repeat_forms(db_data)
}
\arguments{
\item{db_data}{A REDCap dataframe.}
}
\value{
A boolean.
}
\description{
Simple utlity function checking for the existence of repeat forms in a REDCap
database.
}
\keyword{internal}
