% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap.R
\name{distill_nonrepeat_table}
\alias{distill_nonrepeat_table}
\title{Extract Non-Repeat Tables from non-Longitudinal REDCap Databases}
\usage{
distill_nonrepeat_table(form_name, db_data, db_metadata)
}
\arguments{
\item{form_name}{The \code{form_name} described in the named column from the
REDCap metadata.}

\item{db_data}{The REDCap database output defined by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata}{The REDCap metadata output defined by
\code{REDCapR::redcap_metadata_read()$data}}
}
\value{
A \code{tibble} of all data related to a specified \code{form_name}
}
\description{
Sub-helper function to \code{clean_redcap} for single nonrepeat table extraction.
}
\keyword{internal}
