% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap.R
\name{distill_repeat_table}
\alias{distill_repeat_table}
\title{Extract repeat tables from non-longitudinal REDCap databases}
\usage{
distill_repeat_table(form_name, db_data, db_metadata)
}
\arguments{
\item{form_name}{The \code{form_name} described in the named column from the
REDCap metadata.}

\item{db_data}{The non-longitudinal REDCap database output defined by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata}{The non-longitudinal REDCap metadata output defined by
\code{REDCapR::redcap_metadata_read()$data}}
}
\value{
A subset \code{tibble} of all data related to a specified \code{form_name}
}
\description{
Sub-helper function to \code{clean_redcap} for single repeat table
extraction.
}
\keyword{internal}
