% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap_long.R
\name{distill_repeat_table_long}
\alias{distill_repeat_table_long}
\title{Extract repeat tables from longitudinal REDCap databases}
\usage{
distill_repeat_table_long(
  form_name,
  db_data_long,
  db_metadata_long,
  linked_arms,
  has_mixed_structure_forms = FALSE,
  mixed_structure_ref = NULL
)
}
\arguments{
\item{form_name}{The \code{form_name} described in the named column from the
REDCap metadata.}

\item{db_data_long}{The REDCap database output defined by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata_long}{The REDCap metadata output defined by
\code{REDCapR::redcap_metadata_read()$data}}

\item{linked_arms}{Output of \code{link_arms}, linking instruments to REDCap
events/arms}

\item{has_mixed_structure}{Whether the instrument under evaluation has a mixed
structure. Default \code{FALSE}.}

\item{name}{mixed_structure_ref A mixed structure reference dataframe supplied
by \code{get_mixed_structure_fields()}.}
}
\value{
A \code{tibble} of all data related to a specified \code{form_name}
}
\description{
Sub-helper function to \code{clean_redcap_long} for single repeat table
extraction.
}
\keyword{internal}
