% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_tibbles.R
\name{extract_keys}
\alias{extract_keys}
\title{Extract the primary keys associated with a data tibble}
\usage{
extract_keys(data_tbl, record_id_field)
}
\arguments{
\item{data_tbl}{A data tibble from a supertibble}

\item{record_id_field}{The record ID field for the REDCap project, retrieved
as an ouput of \code{\link[=get_record_id_field]{get_record_id_field()}}}
}
\value{
a character string
}
\description{
Extract the primary keys associated with a data tibble
}
\keyword{internal}
