% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_metadata_tibble}
\alias{extract_metadata_tibble}
\title{Extract a specific metadata tibble from a supertibble}
\usage{
extract_metadata_tibble(supertbl, redcap_form_name)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{\link[=read_redcap]{read_redcap()}}.}

\item{redcap_form_name}{A character string identifying the \code{redcap_form_name}
the metadata tibble is associated with.}
}
\value{
A tibble
}
\description{
Utility function to extract a specific metadata tibble from a supertibble
given a \code{redcap_form_name}
}
\keyword{internal}
