% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap_long.R
\name{extract_nonrepeat_table_long}
\alias{extract_nonrepeat_table_long}
\title{Extract Non-Repeat Tables from Longitudinal REDCap Databases}
\usage{
extract_nonrepeat_table_long(
  form_name,
  db_data_long,
  db_metadata_long,
  linked_arms
)
}
\arguments{
\item{form_name}{The \code{form_name} described in the named column from the REDCap metadata.}

\item{linked_arms}{Output of \code{link_arms}, linking forms to REDCap events/arms}

\item{db_data}{The REDCap database output defined by \code{REDCapR::reedcap_read_oneshot()$data}}

\item{db_metadata}{The REDCap metadata output defined by \code{REDCapR::redcap_metadata_read()$data}}
}
\description{
Extract Non-Repeat Tables from Longitudinal REDCap Databases
}
