% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap.R
\name{extract_repeat_table}
\alias{extract_repeat_table}
\title{Extract Repeat Tables from non-Longitudinal REDCap Databases}
\usage{
extract_repeat_table(form_name, db_data, db_metadata)
}
\arguments{
\item{form_name}{The \code{form_name} described in the named column from the REDCap metadata.}

\item{db_data}{The non-longitudinal REDCap database output defined by \code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata}{The non-longitudinal REDCap metadata output defined by \code{REDCapR::redcap_metadata_read()$data}}
}
\description{
Extract Repeat Tables from non-Longitudinal REDCap Databases
}
\keyword{internal}
