% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract a single tibble from a REDCapTidieR supertibble}
\usage{
extract_table(.data, tbl)
}
\arguments{
\item{.data}{A tibble generated by \code{read_redcap_tidy()}. Required.}

\item{tbl}{The \code{form_name} to extract. Required.}
}
\value{
A \code{tibble}
}
\description{
Take a supertibble generated with \code{read_redcap_tidy()}
and return one of its tibbles.
}
\details{
This function makes it easy to extract a single instrument's data from a
REDCapTidieR supertibble.
}
\examples{
\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

out <- read_redcap_tidy(
          redcap_uri,
          token,
          raw_or_label = "label"
        )

 extract_table(out, "sample_form_name")
}

}
