% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{format-helpers}
\alias{format-helpers}
\alias{fmt_strip_whitespace}
\alias{fmt_strip_trailing_colon}
\alias{fmt_strip_trailing_punct}
\alias{fmt_strip_html}
\alias{fmt_strip_field_embedding}
\title{Format REDCap variable labels for display}
\usage{
fmt_strip_whitespace(x)

fmt_strip_trailing_colon(x)

fmt_strip_trailing_punct(x)

fmt_strip_html(x)

fmt_strip_field_embedding(x)
}
\arguments{
\item{x}{a vector of labels}
}
\value{
a vector of formatted labels
}
\description{
Use these functions with the \code{format_labels} argument of
\code{make_labelled()} to define how variable labels are formatted before
being applied to the elements of \code{redcap_data}.

\code{fmt_strip_whitespace()} removes extra whitespace inside and at the ends
of labels. It is a thin wrapper of \code{stringr::str_trim()} and
\code{stringr::str_squish()}. \code{fmt_strip_trailing_colon()} and
\code{fmt_strip_trailing_punct()} remove terminal punctuation.
\code{fmt_strip_html()} removes html tags. \code{fmt_strip_field_embedding()}
removes text between curly braces (\code{\{\}}) which contains special
"field embedding" logic in REDCap. Note that \code{read_redcap_tidy()}
removes html tags and field embedding logic from field labels in the metadata
by default.
}
\examples{

fmt_strip_whitespace("Poorly Spaced   Label ")

fmt_strip_trailing_colon("Label:")

fmt_strip_trailing_punct("Label-")

fmt_strip_html("<b>Bold Label</b>")

fmt_strip_field_embedding("Label{another_field}")

supertbl <- tibble::tribble(
  ~ redcap_data, ~ redcap_metadata,
  tibble::tibble(x = letters[1:3]), tibble::tibble(field_name = "x", field_label = "X Label:")
)

make_labelled(supertbl, format_labels = fmt_strip_trailing_colon)

}
