% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{format_error_val}
\alias{format_error_val}
\title{Format value for error message}
\usage{
format_error_val(x)
}
\arguments{
\item{x}{value to format}
}
\value{
If x is atomic, x with cli formatting to truncate to 5 values. Otherwise,
a string summarizing x produced by as_label
}
\description{
Format value for error message
}
\keyword{internal}
