% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{get_credentials}
\alias{get_credentials}
\title{Retrieve credentials for creating or using mocks}
\usage{
get_credentials(credentials = NULL, fake = FALSE)
}
\arguments{
\item{credentials}{names of credentials to retrieve. By default all credentials in
\code{inst/misc/fake_credentials.csv} are retrieved}

\item{fake}{logical. Should fake credentials be retrieved? By default \code{FALSE}}
}
\value{
A named list of credentials from with \code{Sys.getenv()} if \code{fake = FALSE}
or from \code{inst/misc/fake_credentials.csv} if \code{fake = TRUE}
}
\description{
Retrieve credentials for creating or using mocks
}
\keyword{internal}
