% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_tibbles.R
\name{get_join_fn}
\alias{get_join_fn}
\title{Retrieve the appropriate user specified join function}
\usage{
get_join_fn(type)
}
\arguments{
\item{type}{A character string for the type of join to be performed borrowing from
dplyr. One of "left", "right", "inner", or "full". Default "left".}
}
\value{
a function
}
\description{
Retrieve the appropriate user specified join function
}
\keyword{internal}
