% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_multi_to_single_column.R
\name{get_metadata_ref}
\alias{get_metadata_ref}
\title{Utility function for getting metadata raw and label values for checkboxes}
\usage{
get_metadata_ref(data, supertbl, tbl, instrument_identifiers)
}
\arguments{
\item{data}{a data tibble}

\item{supertbl}{A supertibble generated by \code{\link[=read_redcap]{read_redcap()}}.}

\item{tbl}{The name of the REDCap form (instrument) containing the checkbox
fields.}

\item{instrument_identifiers}{Character string vector of project record identifier vars}
}
\value{
a tibble
}
\description{
Utility function for getting metadata raw and label values for checkboxes
}
\keyword{internal}
