% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_checkboxes.R
\name{get_metadata_spec}
\alias{get_metadata_spec}
\title{Get metadata specification table}
\usage{
get_metadata_spec(metadata_tbl, selected_cols, names_prefix, names_sep)
}
\arguments{
\item{metadata_tbl}{A metadata tibble from the supertibble generated by \code{\link[=read_redcap]{read_redcap()}}.}

\item{selected_cols}{Character string vector of field names for checkbox combination}

\item{names_prefix}{String added to the start of every variable name.}

\item{names_sep}{String to separate new column names from \code{names_prefix}.}
}
\value{
a tibble
}
\description{
Get metadata specification table
}
\keyword{internal}
