% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap_long.R
\name{get_mixed_structure_fields}
\alias{get_mixed_structure_fields}
\title{Get Mixed Structure Instrument List}
\usage{
get_mixed_structure_fields(db_data)
}
\arguments{
\item{db_data}{The REDCap database output generated by
\code{REDCapR::redcap_read_oneshot()$data}}
}
\value{
a dataframe
}
\description{
Define fields in a given project that are used in both a repeating and
nonrepeating manner.
}
\keyword{internal}
