% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_tidy.R
\name{get_output_fields}
\alias{get_output_fields}
\title{Determine the maximal set of fields to keep in the result of read_redcap_tidy}
\usage{
get_output_fields(db_metadata, forms)
}
\arguments{
\item{db_metadata}{metadata tibble created by \code{REDCapR::redcap_metadata_read}}

\item{forms}{\code{forms} argument passed to \code{read_redcap_tidy}}
}
\value{
A character vector fo field names that can be used to filter the results of
\code{REDCapR::redcap_read_oneshot}
}
\description{
Determine the maximal set of fields to keep in the result of read_redcap_tidy
}
\details{
This function applies rules to determine which fields are kept in the results
of calling \code{REDCapR::redcap_read_oneshot}.
}
\keyword{internal}
