% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_record_id_field}
\alias{get_record_id_field}
\title{Utility function to extract the name of the project identifier field for
a tibble of REDCap data}
\usage{
get_record_id_field(data)
}
\arguments{
\item{data}{a tibble of REDCap data}
}
\value{
The name of the identifier field in the data
}
\description{
Utility function to extract the name of the project identifier field for
a tibble of REDCap data
}
\details{
The current implementation assumes that the first field in the data is the
project identifier
}
\keyword{internal}
