% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_tibbles.R
\name{get_structure}
\alias{get_structure}
\title{Retrieve the structure data for a form from the supertibble}
\usage{
get_structure(supertbl, tbl_name)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{read_redcap()}. Required.}

\item{tbl_name}{the \code{x} or \code{y} values assigned to \code{join_data_tibbles}}
}
\description{
Retrieve the structure data for a form from the supertibble
}
\keyword{internal}
