% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data_tibbles.R
\name{get_type}
\alias{get_type}
\title{Retrieve the repeat event type data for a form from the supertibble}
\usage{
get_type(supertbl, tbl_name)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{read_redcap()}. Required.}

\item{tbl_name}{the \code{x} or \code{y} values assigned to \code{join_data_tibbles}}
}
\description{
Retrieve the repeat event type data for a form from the supertibble
}
\keyword{internal}
