% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_vec}
\alias{invert_vec}
\title{Swap vector names for values}
\usage{
invert_vec(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Vector with names and values reversed
}
\description{
Swap vector names for values
}
\keyword{internal}
