% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_labelled}
\alias{is_labelled}
\title{Determine if an object is labelled}
\usage{
is_labelled(obj)
}
\arguments{
\item{obj}{An object to be tested for "label" attributes}
}
\value{
A boolean
}
\description{
An internal utility function used to inform other processes of whether or
not a given object has been labelled (i.e. with \code{make_labelled()}).
}
\details{
An object is considered labelled if it has "label" attributes.
}
\keyword{internal}
