% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_select.R
\name{list_select}
\alias{list_select}
\title{Select Data Tibbles from Named Lists}
\usage{
list_select(list, tbls = everything())
}
\arguments{
\item{list}{A named list from which data tibbles are to be selected. Required.}

\item{tbls}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> The names of the data tibbles to select from the list. Required.}
}
\value{
A named list of selected data tibbles.
}
\description{
\code{list_select()} helps users easily select data tibbles from named lists. This
function uses <\code{\link[=tidyr_tidy_select]{tidy-select}}> syntax to specify which elements to extract.
}
\details{
\code{list_select()} can be used with any named list, and is typically used in
conjunction with \code{\link[=extract_tibbles]{extract_tibbles()}} to pull out named data tibbles of interest
for analytic operations.
}
\examples{
my_list <- extract_tibbles(superheroes_supertbl)

list_select(my_list, starts_with("hero"))

}
