% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{make_labelled}
\alias{make_labelled}
\title{Apply \code{labelled} labels to the elements of a REDCap supertibble based on
project metadata}
\usage{
make_labelled(supertbl, format_labels = NULL)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap_tidy()}}

\item{format_labels}{One of:
\itemize{
\item \code{NULL} to apply variable labels to elements of \code{redcap_data}
as they appear in \code{redcap_metadata}. Default.
\item A function that takes a character vector and returns a modified
character vector of the same length. This function is applied to field
labels before attaching them to variables.
\item A character with the name of a function to be used to format labels
\item A list of label formatting functions or function names to be applied
in order. Note that ordering may affect results.
}}
}
\value{
A labelled supertibble.
}
\description{
This function applies variable labels to the columns of:
\itemize{
\item the supertibble itself
\item each element of \code{redcap_data} based on the \code{field_label} column
in \code{redcap_metadata}
\item each element of \code{redcap_metadata}
\item each element of \code{redcap_events}
}
}
\examples{
supertbl <- tibble::tribble(
  ~ redcap_data, ~ redcap_metadata,
  tibble::tibble(x = letters[1:3]), tibble::tibble(field_name = "x", field_label = "X Label"),
  tibble::tibble(y = letters[1:3]), tibble::tibble(field_name = "y", field_label = "Y Label")
)

make_labelled(supertbl)

make_labelled(supertbl, format_labels = tolower)

\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

supertbl <- read_redcap_tidy(redcap_uri, token)
make_labelled(supertbl)
}
}
