% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_skimr_labels}
\alias{make_skimr_labels}
\title{Make skimr labels from default skimr outputs}
\usage{
make_skimr_labels()
}
\value{
A character vector
}
\description{
A simple helper function that returns all default \code{skimr} names as formatted
character vector for use in \code{make_lablled}
}
\details{
All labels supplied are manually created and agreed upon as human-readable
}
\keyword{internal}
