% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{make_skimr_labs}
\alias{make_skimr_labs}
\title{Make skimr labels from default skimr outputs}
\usage{
make_skimr_labs()
}
\value{
A character vector
}
\description{
Make skimr labels from default skimr outputs
}
\details{
A simple helper function that returns all default \code{skimr} names as formatted
character vector for use in \code{make_lablled}
}
\keyword{internal}
