% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multi_choice_to_labels}
\alias{multi_choice_to_labels}
\title{Update Multiple Choice Fields with Label Data}
\usage{
multi_choice_to_labels(db_data, db_metadata)
}
\arguments{
\item{db_data}{A REDCap database object}

\item{db_metadata}{A REDCap metadata object}
}
\description{
Update Multiple Choice Fields with Label Data
}
\keyword{internal}
