% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_logical_cols}
\alias{parse_logical_cols}
\title{Convert yesno, truefalse, and checkbox fields to logical}
\usage{
parse_logical_cols(db_data, db_metadata, call = caller_env())
}
\arguments{
\item{db_data}{A REDCap database object}

\item{db_metadata}{A REDCap metadata object}

\item{call}{call for conditions}
}
\description{
Convert yesno, truefalse, and checkbox fields to logical
}
\keyword{internal}
