% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_tidy.R
\name{read_redcap_tidy}
\alias{read_redcap_tidy}
\title{Extract REDCap databases to tidy tibbles}
\usage{
read_redcap_tidy(
  redcap_uri,
  token,
  raw_or_label = "label",
  suppress_messages = TRUE
)
}
\arguments{
\item{redcap_uri}{The
uri/url of the REDCap server typically formatted as
"https://server.org/apps/redcap/api/". Required.}

\item{token}{The user-specific string that serves as the password for a
project. Required.}

\item{raw_or_label}{A string (either 'raw' or 'label') that specifies whether
to export the raw coded values or the labels for the options of multiple
choice fields. Default is 'label'.}

\item{suppress_messages}{Optionally show or suppress messages.
Default \code{TRUE}.}
}
\value{
Returns a \code{tibble} with list elements containing tidy dataframes. Users
can access dataframes under \code{$redcap_data} with reference to
\code{form_name} and \code{$structure} column details.
}
\description{
Organize REDCap databases into structured tidy list elements.
}
\details{
\code{REDCapTidieR} will detect REDCap database outputs for structure
elements including:
\itemize{
  \item{Repeat Instruments}
  \item{Longitudinal Events}
  \item{Longitudinal Arms}
}
}
\examples{
\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

read_redcap_tidy(
   redcap_uri,
   token,
   raw_or_label = "label"
 )
}

}
