% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapTidieR-package.R
\docType{package}
\name{REDCapTidieR-package}
\alias{REDCapTidieR-package}
\alias{_PACKAGE}
\title{REDCapTidieR: Extract 'REDCap' Databases into Tidy 'Tibble's}
\description{
Convert 'REDCap' exports into tidy tables for easy handling of 'REDCap' repeat instruments and event arms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chop-cgtinformatics.github.io/REDCapTidieR/}
  \item \url{https://github.com/CHOP-CGTInformatics/REDCapTidieR}
  \item Report bugs at \url{https://github.com/CHOP-CGTInformatics/REDCapTidieR/issues}
}

}
\author{
\strong{Maintainer}: Richard Hanna \email{richardshanna91@gmail.com}

Authors:
\itemize{
  \item Stephan Kadauke \email{kadaukes@chop.edu}
  \item Ezra Porter \email{ezrajporter@gmail.com}
}

}
\keyword{internal}
