% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_multi_to_single_column.R
\name{reduce_multi_to_single_column}
\alias{reduce_multi_to_single_column}
\title{Reduce Multiple Choice Checkbox Fields to Single Column}
\usage{
reduce_multi_to_single_column(
  supertbl,
  tbl,
  cols,
  cols_to,
  multi_val = "Multiple",
  raw_or_label = "label"
)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{\link[=read_redcap]{read_redcap()}}. Required.}

\item{tbl}{The \code{redcap_form_name} of the data tibble under analysis. Required.}

\item{cols}{One or more columns to specify the checkbox field names for conversion. Required.}

\item{cols_to}{A string for the name of the column to consolidate values under. Required.}

\item{raw_or_label}{A string (either 'raw' or 'label') that specifies whether
to export the raw coded values or the labels for the options of categorical
fields. Default is 'label'.}

\item{multi_to}{A string to specify the placeholder value for rows where multiple
checkboxes are selected. Default "Multiple".}
}
\value{
A \code{tibble}.
}
\description{
Convert checkbox fields in a data tibble, represented as many columns, into
a single column with values.
}
