% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{release_questions}
\alias{release_questions}
\title{Additional release questions}
\usage{
release_questions()
}
\value{
A series of character string questions
}
\description{
Additional release questions to be added when using \code{devtools::release()}
during CRAN submissions.
}
\details{
This follows the documentation provided in \code{devtools::release()}.
}
\keyword{internal}
