% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove rows with empty data}
\usage{
remove_empty_rows(data, my_record_id)
}
\arguments{
\item{data}{A REDCap dataframe from a longitudinal database,
pre-processed within a \verb{distill_*} function.}

\item{my_record_id}{The record ID defined in the project.}
}
\value{
A dataframe.
}
\description{
Remove rows that are empty in all associated data columns (those derived from
fields in REDCap). This occurs when a form is filled out in an event, but
other forms are not. Regardless of a form's status, all forms in an event are
included in the output so long as any form in the event contains data.

This only applies to longitudinal REDCap databases containing events.
}
\keyword{internal}
