% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr.R
\name{skim_data}
\alias{skim_data}
\title{Apply applicable skimmers to data}
\usage{
skim_data(redcap_data, redcap_metadata, is_labelled)
}
\value{
A dataframe
}
\description{
A helper function for \code{add_skimr_metadata()} which applies applicable
skimmers to a given dataframe.
}
\keyword{internal}
