% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_html_field_embedding}
\alias{strip_html_field_embedding}
\title{Remove html tags and field embedding logic from a string}
\usage{
strip_html_field_embedding(x)
}
\arguments{
\item{x}{vector of strings to format}
}
\value{
vector of strings with html tags, field embedding logic, and extra whitespace
removed
}
\description{
Remove html tags and field embedding logic from a string
}
