% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{structure_from_events_and_repeating}
\alias{structure_from_events_and_repeating}
\title{Determine form-level structure for a longitudinal REDCap project}
\usage{
structure_from_events_and_repeating(
  forms,
  db_instrument_repeating,
  db_event_instruments
)
}
\arguments{
\item{forms}{a character vector of form names.}

\item{db_instrument_repeating}{a tibble of repeating instrument settings
created by \code{\link[=pull_instrument_repeating]{pull_instrument_repeating()}}.}

\item{db_event_instruments}{a tibble of event-instrument mappings created by
\code{\link[=pull_event_instruments]{pull_event_instruments()}}.}
}
\value{
Tibble with \code{redcap_form_name} and \code{structure}
}
\description{
Determine form-level structure for a longitudinal REDCap project
}
\details{
\code{structure} is one of 'repeating', 'nonrepeating', or 'mixed' based on form-event-level
structure. In cases where a form in \code{forms} is not \code{db_event_instruments}
it will be reported as 'nonrepeating'.
}
\keyword{internal}
