% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{structure_from_repeating}
\alias{structure_from_repeating}
\title{Determine instrument structure for a classic REDCap project}
\usage{
structure_from_repeating(form_structure, db_instrument_repeating)
}
\arguments{
\item{form_structure}{a tibble containing one row per instrument and a
\code{structure} column.}

\item{db_instrument_repeating}{a tibble of repeating instrument settings
created by \code{\link[=pull_instrument_repeating]{pull_instrument_repeating()}}.}
}
\value{
The original \code{form_structure} tibble with \code{structure} updated to
'repeating' if present in \code{db_instrument_repeating}
}
\description{
Determine instrument structure for a classic REDCap project
}
\keyword{internal}
