% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superheroes_supertbl}
\alias{superheroes_supertbl}
\title{Superheroes Data}
\format{
\subsection{\code{heroes_information}}{

A \code{tibble} with 734 rows and 12 columns:
\describe{
\item{record_id}{REDCap record ID}
\item{name}{Hero name}
\item{gender}{Gender}
\item{eye_color}{Eye color}
\item{race}{Race}
\item{hair_color}{Hair color}
\item{height}{Height}
\item{weight}{Weight}
\item{publisher}{Publisher}
\item{skin_color}{Skin color}
\item{alignment}{Alignment}
\item{form_status_complete}{REDCap instrument completed?}
}
}

\subsection{\code{super_hero_powers}}{

A \code{tibble} with 5,966 rows and 4 columns:
\describe{
\item{record_id}{REDCap record ID}
\item{redcap_repeat_instance}{REDCap repeat instance}
\item{power}{Super power}
\item{form_status_complete}{REDCap instrument completed?}
}
}
}
\source{
\url{https://www.superherodb.com/}
}
\usage{
superheroes_supertbl
}
\description{
A dataset of superheroes in a REDCapTidieR \code{supertbl} object
}
\keyword{datasets}
