% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{supertbl_recode}
\alias{supertbl_recode}
\title{Recode fields using supertbl metadata}
\usage{
supertbl_recode(supertbl, supertbl_meta, recode_yn)
}
\arguments{
\item{supertbl}{A supertibble generated using \code{read_redcap()}}

\item{recode_yn}{Convert REDCap yesno questions from TRUE/FALSE to "yes"/"no"
for readability. Default TRUE.}

\item{superbl_meta}{an \code{unnest}-ed metadata tibble from the supertibble}
}
\description{
This utility function helps to map metadata field types in order to apply
changes in supertbl tables.
}
\keyword{internal}
