% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertibble.R
\name{tbl_sum.redcap_supertbl}
\alias{tbl_sum.redcap_supertbl}
\title{Provide a succinct summary of an object}
\usage{
\method{tbl_sum}{redcap_supertbl}(x)
}
\arguments{
\item{x}{Object to summarise.}
}
\value{
A named character vector, describing the dimensions in the first element
and the data source in the name of the first element.
}
\description{
\code{tbl_sum()} gives a brief textual description of a table-like object,
which should include the dimensions and the data source in the first element,
and additional information in the other elements (such as grouping for \pkg{dplyr}).
The default implementation forwards to \code{\link[pillar:obj_sum]{obj_sum()}}.
}
