% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_data_col_names}
\alias{update_data_col_names}
\title{Database Colname Helper Function for Checkboxes with Minus Signs}
\usage{
update_data_col_names(db_data, db_metadata)
}
\arguments{
\item{db_data}{A REDCap database object}

\item{db_metadata}{A REDCap metadata object}
}
\value{
Returns an updated REDCap database object with updated checkbox column names
}
\description{
Takes a \code{db_metadata} object and appends a \code{field_name_updated} field to the end for checkbox variable handling.
}
\keyword{internal}
