% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_field_names}
\alias{update_field_names}
\title{Metadata field_name Updater Helper Function}
\usage{
update_field_names(db_metadata)
}
\arguments{
\item{db_metadata}{A REDCap metadata object}
}
\description{
Takes a \code{db_metadata} object and appends a \code{field_name_updated} field to the end for checkbox variable handling.
}
\keyword{internal}
