% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertibble.R
\name{vec_ptype_abbr.redcap_supertbl}
\alias{vec_ptype_abbr.redcap_supertbl}
\title{Vector type as a string}
\usage{
\method{vec_ptype_abbr}{redcap_supertbl}(x, ..., prefix_named, suffix_shape)
}
\arguments{
\item{x}{A vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{prefix_named}{If \code{TRUE}, add a prefix for named vectors.}

\item{suffix_shape}{If \code{TRUE} (the default), append the shape of
the vector.}
}
\value{
A string.
}
\description{
\code{vec_ptype_full()} displays the full type of the vector. \code{vec_ptype_abbr()}
provides an abbreviated summary suitable for use in a column heading.
}
