% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{add_event_mapping}
\alias{add_event_mapping}
\title{Supplement a supertibble from a longitudinal database with information about
the events associated with each instrument}
\usage{
add_event_mapping(supertbl, linked_arms)
}
\arguments{
\item{supertbl}{a supertibble object to supplement with metadata}

\item{linked_arms}{the tibble with event mappings created by
\code{link_arms()}}
}
\value{
The original supertibble with an events \code{redcap_events} list column
containing arms and events associated with each instrument
}
\description{
Supplement a supertibble from a longitudinal database with information about
the events associated with each instrument
}
\keyword{internal}
