% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{add_labelled_xlsx_features}
\alias{add_labelled_xlsx_features}
\title{Add labelled features to write_supertibble_xlsx}
\usage{
add_labelled_xlsx_features(
  supertbl,
  wb,
  sheet_vals,
  incl_supertbl = TRUE,
  incl_meta = TRUE,
  supertbl_toc
)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{wb}{An \code{openxlsx} workbook object}

\item{sheet_vals}{Helper argument passed from \code{write_supertibble_xlsx} to
determine and assign sheet values.}

\item{incl_supertbl}{Include a sheet capturing the supertibble output.
Default \code{TRUE}.}

\item{incl_meta}{Include a sheet capturing the combined output of the
supertibble \code{redcap_metadata}. Default \code{TRUE}.}

\item{supertbl_toc}{The table of contents supertibble defined in the parent
function.}
}
\description{
Helper function to support \code{labelled} aesthetics to XLSX supertibble output
}
\keyword{internal}
