% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{add_labelled_xlsx_features}
\alias{add_labelled_xlsx_features}
\title{Add labelled features to write_supertibble_xlsx}
\usage{
add_labelled_xlsx_features(supertbl, wb)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{wb}{An \code{openxlsx} workbook object}
}
\description{
Helper function to support \code{labelled} aesthetics to XLSX supertibble output
}
\keyword{internal}
