% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{add_labelled_xlsx_features}
\alias{add_labelled_xlsx_features}
\title{Add labelled features to write_redcap_xlsx}
\usage{
add_labelled_xlsx_features(
  supertbl,
  wb,
  sheet_vals,
  include_toc_sheet = TRUE,
  include_metadata_sheet = TRUE,
  supertbl_toc = NULL
)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{wb}{An \code{openxlsx2} workbook object}

\item{sheet_vals}{Helper argument passed from \code{write_redcap_xlsx} to
determine and assign sheet values.}

\item{include_toc_sheet}{Include a sheet capturing the supertibble output.
Default \code{TRUE}.}

\item{include_metadata_sheet}{Include a sheet capturing the combined output of the
supertibble \code{redcap_metadata}. Default \code{TRUE}.}

\item{supertbl_toc}{The table of contents supertibble defined in the parent
function. Default \code{NULL}.}
}
\description{
Helper function to support \code{labelled} aesthetics to XLSX supertibble output
}
\keyword{internal}
