% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{add_metadata_sheet}
\alias{add_metadata_sheet}
\title{Add the metadata sheet}
\usage{
add_metadata_sheet(
  supertbl,
  supertbl_meta,
  include_metadata,
  wb,
  labelled,
  table_style,
  set_col_widths,
  na_strings
)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{include_metadata}{Include a sheet capturing the combined output of the
supertibble \code{redcap_metadata}. Default \code{TRUE}.}

\item{wb}{An \code{openxlsx2} workbook object}

\item{labelled}{Whether or not to include labelled outputs, default \code{FALSE}.
Requires use of \code{make_labelled()}.}

\item{table_style}{Any excel table style name or "none" (see "formatting"
vignette in \link[openxlsx2]{wb_add_data_table}). Default "tableStyleLight8".}

\item{set_col_widths}{Width to set columns across the workbook. Default
"auto", otherwise a numeric value. Standard Excel is 8.43.}

\item{na_strings}{Value used for replacing NA values from x. Default
\code{na_strings()} uses the special ⁠#N/A⁠ value within the workbook.}

\item{superbl_meta}{an \code{unnest}-ed metadata tibble from the supertibble}
}
\value{
A dataframe
}
\description{
Internal helper function. Adds appropriate elements to wb object. Returns
a dataframe.
}
\keyword{internal}
