% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_partial_keys}
\alias{add_partial_keys}
\title{Add partial key helper variables to dataframes}
\usage{
add_partial_keys(db_data, var)
}
\arguments{
\item{db_data}{The REDCap database output defined by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{var}{the unquoted name of the field containing event and arm
identifiers}
}
\value{
Two appended columns, \code{redcap_event} and \code{redcap_arm}
to the end of \code{read_redcap_tidy} output \code{tibble}s.
}
\description{
Make helper variables \code{redcap_event} and \code{redcap_arm} available as
branches from \code{var} for later use.
}
\keyword{internal}
