% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tables.R
\name{bind_tables}
\alias{bind_tables}
\title{Extract tibbles from a REDCapTidieR supertibble and bind them to an
environment}
\usage{
bind_tables(.data, environment = global_env(), tbls = NULL, structure = NULL)
}
\arguments{
\item{.data}{A supertibble generated by \code{read_redcap_tidy()}. Required.}

\item{environment}{The environment to bind the tibbles to. Default is
\code{rlang::global_env()}.}

\item{tbls}{A vector of the \code{form_name}s of the data tibbles to bind to
the environment. Default is \code{NULL} which binds all data tibbles.}

\item{structure}{A one-element character vector to control whether to extract
only \code{repeating} or \code{nonrepeating} instruments. Default is \code{NULL} which
extracts both types of instruments.}
}
\value{
This function returns nothing as it's used solely for its side effect of
modifying an environment.
}
\description{
Take a supertibble generated with \code{read_redcap_tidy()}
and bind its data tibbles (i.e. the tibbles in the \code{redcap_data} column) to
an environment. The default is the global environment.
}
\examples{
# Designate a sample environment
sample_env <- environment()

# Supply a `read_redcap_tidy()` output to bind
sample_data <- tibble::tribble(
  ~redcap_form_name,    ~redcap_data,   ~structure,
  "super_hero_powers",   list(),         "repeating",
  "heroes_information",  list(),         "nonrepeating"
)

bind_tables(sample_data, sample_env)

}
