% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_extra_field_values}
\alias{check_extra_field_values}
\title{Check data field for field values not in metadata}
\usage{
check_extra_field_values(x, values, field_name, call = caller_env())
}
\arguments{
\item{x}{data field}

\item{field_name}{field name for warning message}

\item{call}{call for warning message}
}
\description{
Check data field for field values not in metadata
}
\keyword{internal}
