% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_field_is_logical}
\alias{check_field_is_logical}
\title{Parse logical field and compile data for warning if parsing errors occurred}
\usage{
check_field_is_logical(x)
}
\arguments{
\item{x}{vector to parse}
}
\description{
Parse logical field and compile data for warning if parsing errors occurred
}
\keyword{internal}
