% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_forms_exist}
\alias{check_forms_exist}
\title{Check that all requested forms are in REDCap project metadata}
\usage{
check_forms_exist(db_metadata, forms)
}
\arguments{
\item{db_metadata}{The metadata file read by
\code{REDCapR::redcap_metadata_read()}}

\item{forms}{The character vector of form names passed to
\code{read_redcap_tidy()}}
}
\value{
An error message listing the requested forms that don't exist
}
\description{
Provide an error message when any form names are passed to
\code{read_redcap_tidy()} that do not exist in the project metadata.
}
\keyword{internal}
