% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{check_labelled}
\alias{check_labelled}
\title{Check if labelled}
\usage{
check_labelled(supertbl, labelled, call = caller_env())
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{labelled}{Whether or not to include labelled outputs, default \code{FALSE}.
Requires use of \code{make_labelled()}.}

\item{call}{the calling environment to use in the warning message}
}
\value{
A boolean
}
\description{
Checks if a supplied supertibble is labelled and throws an error if it is not
but \code{labelled} is set to \code{TRUE}
}
\keyword{internal}
