% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_redcap_populated}
\alias{check_redcap_populated}
\title{Check that a supplied REDCap database is populated}
\usage{
check_redcap_populated(db_data, call = caller_env())
}
\arguments{
\item{db_data}{The REDCap database output generated by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{call}{the calling environment to use in the error message}
}
\value{
A helpful error message alerting the user to check their API privileges.
}
\description{
Check for potential outputs where metadata is present, but \code{nrow} and
\code{ncol} equal \code{0}. This causes \code{multi_choice_to_labels} to fail, but
a helpful error message should be provided.
}
\keyword{internal}
