% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_repeat_and_nonrepeat}
\alias{check_repeat_and_nonrepeat}
\title{Check for Instruments that are Both Repeating and NonRepeating}
\usage{
check_repeat_and_nonrepeat(db_data)
}
\arguments{
\item{db_data}{The REDCap database output generated by \code{REDCapR::redcap_read_oneshot()$data}}
}
\description{
Check for potential instruments that are given both repeating and
nonrepeating structure. This should never be used with \code{REDCapTidieR}.
}
\keyword{internal}
