% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_req_labelled_fields}
\alias{check_req_labelled_fields}
\title{Check that a supertibble contains \code{redcap_data} and
\code{redcap_metadata} fields}
\usage{
check_req_labelled_fields(supertbl)
}
\arguments{
\item{supertbl}{a supertibble}
}
\value{
An error message indicating that the required columns are missing
}
\description{
Provide an error message when a tibble is missing \code{redcap_data} or
\code{redcap_metadata}
}
\keyword{internal}
