% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_user_rights}
\alias{check_user_rights}
\title{Check for possible API user privilege issues}
\usage{
check_user_rights(db_data, db_metadata, call = caller_env())
}
\arguments{
\item{db_data}{The REDCap database output generated by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata}{The REDCap metadata output generated by \code{REDCapR::redcap_metadata_read()$data}}

\item{call}{the calling environment to use in the warning message}
}
\value{
A helpful error message alerting the user to check their API privileges.
}
\description{
Check for potential user access privilege issues and provide an appropriate
warning message. This can occur when metadata forms/field names do not appear
in a database export.
}
\keyword{internal}
