% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkbox_appender}
\alias{checkbox_appender}
\title{Checkbox String Appender Helper Function}
\usage{
checkbox_appender(field_name, string)
}
\arguments{
\item{field_name}{The \code{db_metadata$field_name} to append onto the string}

\item{string}{A \code{db_metadata$select_choices_or_calculations} field
pre-filtered for checkbox \code{field_type}}
}
\value{
A vector.
}
\description{
Takes a \code{db_metadata$select_choices_or_calculations} field pre-filtered
for checkbox \code{field_type}s and returns a vector of key names for
appending to checkbox variables.
}
\keyword{internal}
