% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap.R
\name{clean_redcap}
\alias{clean_redcap}
\title{Extract non-longitudinal REDCap databases into tidy tibbles}
\usage{
clean_redcap(db_data, db_metadata)
}
\arguments{
\item{db_data}{The REDCap database output defined by
\code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata}{The REDCap metadata output defined by
\code{REDCapR::redcap_metadata_read()$data}}
}
\value{
Returns a \code{tibble} with list elements containing tidy dataframes. Users
can access dataframes under the \code{redcap_data} column with reference to
\code{form_name} and \code{structure} column details.
}
\description{
Helper function internal to \code{read_redcap} responsible for
extraction and final processing of a tidy \code{tibble} to the user from
a non-longitudinal REDCap database.
}
\keyword{internal}
