% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_redcap_long.R
\name{clean_redcap_long}
\alias{clean_redcap_long}
\title{Extract Longitudinal REDCap Databases into Tidy Tibbles}
\usage{
clean_redcap_long(db_data_long, db_metadata_long, linked_arms, has_repeating)
}
\arguments{
\item{db_data_long}{The longitudinal REDCap database output defined by \code{REDCapR::redcap_read_oneshot()$data}}

\item{db_metadata_long}{The longitudinal REDCap metadata output defined by \code{REDCapR::redcap_metadata_read()$data}}

\item{linked_arms}{Output of \code{link_arms}, linking forms to REDCap events/arms}

\item{has_repeating}{T/F, does the supplied REDCap database contain repeating instruments as determined by its contents.}
}
\description{
Extract Longitudinal REDCap Databases into Tidy Tibbles
}
\keyword{internal}
