% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_checkboxes.R
\name{combine_checkboxes}
\alias{combine_checkboxes}
\title{Combine Checkbox Fields into a Single Column}
\usage{
combine_checkboxes(
  supertbl,
  tbl,
  cols,
  names_prefix = "",
  names_sep = "_",
  names_glue = NULL,
  names_repair = "check_unique",
  multi_value_label = "Multiple",
  values_fill = NA,
  raw_or_label = "label",
  keep = TRUE
)
}
\arguments{
\item{supertbl}{A supertibble generated by \code{\link[=read_redcap]{read_redcap()}}. Required.}

\item{tbl}{The \code{redcap_form_name} of the data tibble to extract. Required.}

\item{cols}{<\code{\link[=tidyr_tidy_select]{tidy-select}}> Checkbox columns to combine to
single column. Required.}

\item{names_prefix}{String added to the start of every variable name.}

\item{names_sep}{String to separate new column names from \code{names_prefix}.}

\item{multi_value_label}{A string specifying the value to be used when multiple
checkbox fields are selected. Default "Multiple".}

\item{values_fill}{Value to use when no checkboxes are selected. Default \code{NA}.}

\item{raw_or_label}{Either 'raw' or 'label' to specify whether to use raw coded
values or labels for the options. Default 'label'.}

\item{keep}{Logical indicating whether to keep the original checkbox fields in
the output. Default \code{TRUE}.}
}
\value{
A modified supertibble.
}
\description{
\code{\link[=combine_checkboxes]{combine_checkboxes()}} consolidates multiple checkbox fields in a REDCap data
tibble into a single column. This transformation simplifies analysis by
combining multiple binary columns into a singular and informative labelled
factor column.
}
\details{
\code{\link[=combine_checkboxes]{combine_checkboxes()}} makes use of the output names of \code{\link[=read_redcap]{read_redcap()}}
data tibbles and metadata tibbles. Changes to checkbox data names or
metadata \code{field_name}s may result in errors.

Checkbox fields are expanded to be a variable per checkbox option, separated
by underscores. For example, \code{checkbox_var} with 2 options becomes
\code{checkbox_var___1} and \code{checkbox_var___2}. \code{\link[=combine_checkboxes]{combine_checkboxes()}} looks for
these and may give a error if it cannot find them.
}
\examples{
\dontrun{
supertbl <- read_redcap(redcap_uri, token)
combine_checkboxes(
  supertbl = supertbl,
  tbl = "demographics",
  cols = starts_with("race"),
  multi_value_label = "Multiple",
  values_fill = NA
)
}

}
