% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract a Single Specific REDCapTidieR Table}
\usage{
extract_table(.data, tbl)
}
\arguments{
\item{.data}{A tidy table provided by \code{read_redcap_tidy()}}

\item{tbl}{REDCap table name specification, one max.}
}
\description{
Supply a \code{read_redcap_tidy()} output and specify a table of interest to extract.
}
