% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract one or more dataframes from a REDCapTidieR output}
\usage{
extract_tables(.data, tbls = everything())
}
\arguments{
\item{.data}{A REDCapTidieR output provided by \code{read_redcap_tidy()}}

\item{tbls}{REDCap \code{form_name} specification, one or more. Default is
\code{everything()}.}
}
\value{
A named list of \code{tibble}s specified by the user from specified
\code{form_name}s.
}
\description{
Given a \code{read_redcap_tidy()} output and one or more \code{form_name}s,
return a named list back to the user.
}
\details{
Users may supply \code{tidyselect} statements such as \code{starts_with()}
or \code{ends_with()} for easier selection.
}
\examples{
\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

out <- read_redcap_tidy(
          redcap_uri,
          token,
          raw_or_label = "label"
        )

 extract_tables(out, c("sample_form_name1", "sample_form_name2"))
 extract_tables(out, everything())
 extract_tables(out, starts_with("demographics"))
}

}
