% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract Tables from a REDCapTidieR Supertibble}
\usage{
extract_tables(.data, tbls)
}
\arguments{
\item{.data}{A tidy table provided by \code{read_redcap_tidy()}}

\item{tbls}{REDCap table name specifications, one or more.}
}
\value{
A named list of \code{tibble}s specified by the user.
}
\description{
Supply a \code{read_redcap_tidy()} output and specify tables of interest to extract.
Users may supply \code{tidyselect} statements such as `starts_with()` or `ends_with()` for easier selection.
}
