% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_checkboxes.R
\name{get_metadata_ref}
\alias{get_metadata_ref}
\title{Utility function for getting metadata raw and label values for checkboxes}
\usage{
get_metadata_ref(data, supertbl, tbl, field_names)
}
\arguments{
\item{data}{a data tibble}

\item{supertbl}{A supertibble generated by \code{\link[=read_redcap]{read_redcap()}}.}

\item{tbl}{The name of the REDCap form (instrument) containing the checkbox
fields.}

\item{field_names}{Character string vector of field names for checkbox combination}
}
\value{
a tibble
}
\description{
Utility function for getting metadata raw and label values for checkboxes
}
\keyword{internal}
