% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_checkboxes.R
\name{get_metadata_ref}
\alias{get_metadata_ref}
\title{Utility function for getting metadata raw and label values for checkboxes}
\usage{
get_metadata_ref(metadata_tbl, selected_cols)
}
\arguments{
\item{metadata_tbl}{A metadata tibble from the supertibble generated by \code{\link[=read_redcap]{read_redcap()}}.}

\item{selected_cols}{Character string vector of field names for checkbox combination}
}
\value{
a tibble
}
\description{
Utility function for getting metadata raw and label values for checkboxes
}
\keyword{internal}
