% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link_arms}
\alias{link_arms}
\title{Link longitudinal REDCap instruments with their events/arms}
\usage{
link_arms(redcap_uri, token, suppress_redcapr_messages = TRUE)
}
\arguments{
\item{redcap_uri}{The REDCap URI}

\item{token}{The REDCap API token}

\item{suppress_redcapr_messages}{A logical to control whether to suppress messages
from REDCapR API calls. Default \code{TRUE}.}
}
\value{
Returns a \code{tibble} of \code{redcap_event_name}s with list elements
containing a vector of associated instruments.
}
\description{
For REDCap databases containing arms and events, it is necessary to determine
how these are linked and what variables belong to them.
}
\keyword{internal}
