% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link_arms}
\alias{link_arms}
\title{Link Longitudinal REDCap Forms with the Events/Arms They Belong To}
\usage{
link_arms(db_data_long, db_metadata_long, redcap_uri, token)
}
\arguments{
\item{db_data_long}{A longitudinal REDCap database export}

\item{db_metadata_long}{A longitudinal REDCap metadata export}

\item{redcap_uri}{The REDCap URI}

\item{token}{The REDCap API token}
}
\value{
Returns a \code{tibble} of \code{redcap_event_name}s with list elements
containing a vector of associated forms.
}
\description{
For REDCap databases containing arms and events, it is necessary to determine
how these are linked and what variables belong to them.
}
\keyword{internal}
