% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_labels}
\alias{parse_labels}
\title{Parse labels from REDCap metadata into usable formats}
\usage{
parse_labels(string, return_vector = FALSE)
}
\arguments{
\item{string}{A \code{db_metadata$select_choices_or_calculations} field
pre-filtered for checkbox \code{field_type}}

\item{return_vector}{logical for whether to return result as a vector}
}
\value{
A tidy \code{tibble} from a matrix giving raw and label outputs to
be used in later functions if \code{return_vector = FALSE}, the default.
Otherwise a vector result in a c(raw = label) format to use with
dplyr::recode
}
\description{
Takes a string separated by \code{,}s and/or \code{|}s (i.e. comma/tab
separated values) containing key value pairs (\code{raw} and \code{label})
and returns a tidy \code{tibble}.
}
\details{
The associated \code{string} comes from metadata outputs.
}
\keyword{internal}
