% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{pull_instrument_repeating}
\alias{pull_instrument_repeating}
\title{Retrieve repeating instrument settings from REDCap}
\usage{
pull_instrument_repeating(
  redcap_uri,
  token,
  suppress_redcapr_messages = TRUE,
  call = NULL
)
}
\arguments{
\item{redcap_uri}{The
URI/URL of the REDCap server (e.g.,
"https://server.org/apps/redcap/api/"). Required.}

\item{token}{The user-specific string that serves as the password for a
project. Required.}

\item{suppress_redcapr_messages}{A logical to control whether to suppress messages
from REDCapR API calls. Default \code{TRUE}.}

\item{call}{The calling environment to use when handling REDCapR errors.
Default \code{NULL}.}
}
\value{
A tibble of repeating instrument settings with standardized column names,
including \code{unique_event_name} and \code{form} when available.
}
\description{
Retrieve repeating instrument settings from REDCap
}
\keyword{internal}
