% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_tidy.R
\name{read_redcap_tidy}
\alias{read_redcap_tidy}
\title{Extract REDCap Databases to Tidy Tibbles}
\usage{
read_redcap_tidy(redcap_uri, token)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap project. Required.}

\item{token}{The user-specific string that serves as the password for a project. Required.}
}
\description{
Automatically detect REDCap database structure for the following elements:
\itemize{
  \item{Repeat Instruments}
  \item{Longitudinal Events}
  \item{Longitudinal Arms}
}
Output a \code{tibble} with list elements containing tidy dataframes. Ideal for combination of tables with join operations primary and composite keys.
}
