% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_tidy.R
\name{read_redcap_tidy}
\alias{read_redcap_tidy}
\title{Extract a REDCap database into a tidy supertibble}
\usage{
read_redcap_tidy(
  redcap_uri,
  token,
  raw_or_label = "label",
  forms = NULL,
  export_survey_fields = TRUE,
  suppress_messages = TRUE
)
}
\arguments{
\item{redcap_uri}{The
URI/URL of the REDCap server (e.g.,
"https://server.org/apps/redcap/api/"). Required.}

\item{token}{The user-specific string that serves as the password for a
project. Required.}

\item{raw_or_label}{A string (either 'raw' or 'label') that specifies whether
to export the raw coded values or the labels for the options of categorical
fields. Default is 'label'.}

\item{forms}{A character vector of form names that specifies the forms to
export. Default is \code{NULL} which returns all forms in the project.}

\item{export_survey_fields}{A logical that specifies whether to export the
survey identifier field (e.g., 'redcap_survey_identifier') or survey
timestamp fields ('[instrument_name]_timestamp'). The timestamp data
reflect the survey's completion time (according to the time and timezone of
the REDCap server.). Default is \code{TRUE}.}

\item{suppress_messages}{A logical to control whether to suppress messages
from REDCapR API calls. Default \code{TRUE}.}
}
\value{
Returns a \code{tibble} in which each row represents a REDCap instrument. The
results contains the following fields:
* \code{redcap_form_name}, the name of the instrument
* \code{redcap_form_label}, the label for the instrument
* \code{redcap_data}, a \code{tibble} with the data for the instrument
* \code{redcap_metadata}, a \code{tibble} of data dictionary entries for each field in the instrument
* \code{redcap_events}, a \code{tibble} with information about the arms and longitudinal events represented in the instrument. Only if the project has longitudinal events enabled
* \code{structure}, either "repeating" or "nonrepeating"
* \code{data_rows}, the number of rows in the instrument's data tibble
* \code{data_cols}, the number of columns in the instrument's data tibble
* \code{data_size}, the size in memory of the instrument's data tibble computed by \code{lobstr::obj_size}
* \code{data_na_pct}, the percentage of cells in the instrument's data that are \code{NA} excluding identifier and form completion fields
}
\description{
Call the REDCap API to retrieve data and metadata about a project,
and then transform the output into a tidy "supertibble" that contains data
and metadata organized into tidy tibbles, broken down by instrument.
}
\details{
This function uses the \link[=https://ouhscbbmc.github.io/REDCapR/index.html]{REDCapR}
package to call the REDCap API. The REDCap API returns a
\link[https://en.wikipedia.org/wiki/Block_matrix]{block matrix} that mashes
data from all data collection instruments
together. The `read_redcap_tidy` function intelligently
deconvolutes the block matrix and splices the data into individual tibbles,
where one tibble represents the data from one instrument.
}
\examples{
\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

read_redcap_tidy(
   redcap_uri,
   token,
   raw_or_label = "label"
 )
}

}
