% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{structure_from_events_and_repeating}
\alias{structure_from_events_and_repeating}
\title{Determine form-level structure for a longitudinal REDCap project}
\usage{
structure_from_events_and_repeating(
  form_structure,
  db_instrument_repeating,
  db_event_instruments
)
}
\arguments{
\item{form_structure}{a tibble containing one row per instrument and a
\code{structure} column.}

\item{db_instrument_repeating}{a tibble of repeating instrument settings
created by \code{\link[=pull_instrument_repeating]{pull_instrument_repeating()}}.}

\item{db_event_instruments}{a tibble of event-instrument mappings created by
\code{\link[=pull_event_instruments]{pull_event_instruments()}}.}
}
\value{
The original \code{form_structure} tibble with \code{structure} updated to
'repeating', 'nonrepeating', or 'mixed' based on form-event-level structure.
}
\description{
Determine form-level structure for a longitudinal REDCap project
}
\keyword{internal}
