% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{structure_from_repeating}
\alias{structure_from_repeating}
\title{Determine instrument structure for a classic REDCap project}
\usage{
structure_from_repeating(forms, db_instrument_repeating)
}
\arguments{
\item{forms}{a character vector of form names.}

\item{db_instrument_repeating}{a tibble of repeating instrument settings
created by \code{\link[=pull_instrument_repeating]{pull_instrument_repeating()}}.}
}
\value{
Tibble with \code{redcap_form_name} and \code{structure}
}
\description{
Determine instrument structure for a classic REDCap project
}
\keyword{internal}
