% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{supertbl_recode}
\alias{supertbl_recode}
\title{Recode fields using supertbl metadata}
\usage{
supertbl_recode(supertbl, supertbl_meta, add_labelled_column_headers)
}
\arguments{
\item{supertbl}{A supertibble generated using \code{read_redcap()}}

\item{supertbl_meta}{an \code{unnest}-ed metadata tibble from the supertibble}

\item{add_labelled_column_headers}{Whether or not to include labelled outputs}
}
\description{
This utility function helps to map metadata field types in order to apply
changes in supertbl tables.
}
\keyword{internal}
