% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{update_dag_cols}
\alias{update_dag_cols}
\title{Implement REDCapR DAG Data into Supertibble}
\usage{
update_dag_cols(
  data,
  redcap_uri = NULL,
  token = NULL,
  suppress_redcapr_messages = NULL,
  test_data = FALSE
)
}
\arguments{
\item{data}{the REDCap data}

\item{redcap_uri}{The
URI/URL of the REDCap server (e.g.,
"https://server.org/apps/redcap/api/"). Required.}

\item{token}{The user-specific string that serves as the password for a
project. Required.}

\item{suppress_redcapr_messages}{A logical to control whether to suppress messages
from REDCapR API calls. Default \code{TRUE}.}

\item{test_data}{T/F. Solely used for testing purposes, default \code{FALSE}.}
}
\description{
This helper function calls \link{redcap_dag_read} and applies the necessary
raw/label values to the \code{redcap_data_access_group} column.

This is done because REDCapTidieR retrieves raw data by default, then merges
labels from the metadata. However, some columns like
\code{redcap_data_access_group} are not in the metadata and so there is nothing by
default to reference.
}
\keyword{internal}
