% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_field_names}
\alias{update_field_names}
\title{Update metadata field names for checkbox handling}
\usage{
update_field_names(db_metadata)
}
\arguments{
\item{db_metadata}{The REDCap metadata output defined by
\code{REDCapR::redcap_metadata_read()$data}}
}
\value{
Column \code{field_name_updated} appended to a given REDCap
\code{db_metadata} object.

Returns an updated REDCap metadata object with updated field names
}
\description{
Takes a \code{db_metadata} object and appends a \code{field_name_updated}
field to the end for checkbox variable handling.
}
\keyword{internal}
