% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_redcap_xlsx}
\alias{write_redcap_xlsx}
\title{Write Supertibbles to XLSX}
\usage{
write_redcap_xlsx(
  supertbl,
  file,
  labelled = NULL,
  use_labels_for_sheet_names = TRUE,
  include_toc_from_supertbl = TRUE,
  include_metadata = TRUE,
  table_style = "tableStyleLight8",
  set_col_widths = "auto",
  recode_yn = TRUE,
  na_strings = openxlsx2::na_strings()
)
}
\arguments{
\item{supertbl}{A supertibble generated using \code{read_redcap()}}

\item{file}{A character string naming an xlsx file}

\item{labelled}{Whether or not to include labelled outputs, default \code{NULL}.
Requires use of \code{make_labelled()}.}

\item{use_labels_for_sheet_names}{Whether or not to include labels in the XLSX sheets
instead of raw values. Default \code{TRUE}.}

\item{include_toc_from_supertbl}{Include a sheet capturing the supertibble output.
Default \code{TRUE}.}

\item{include_metadata}{Include a sheet capturing the combined output of the
supertibble \code{redcap_metadata}. Default \code{TRUE}.}

\item{table_style}{Any excel table style name or "none" (see "formatting"
vignette in \link[openxlsx2]{wb_add_data_table}). Default "tableStyleLight8".}

\item{set_col_widths}{Width to set columns across the workbook. Default
"auto", otherwise a numeric value. Standard Excel is 8.43.}

\item{recode_yn}{Convert REDCap yesno questions from TRUE/FALSE to "yes"/"no"
for readability. Default TRUE.}

\item{na_strings}{Value used for replacing NA values from x. Default
\code{na_strings()} uses the special ⁠#N/A⁠ value within the workbook.}
}
\value{
An \code{openxlsx2} workbook object
}
\description{
Write supertibble outputs to XLSX file outputs, where each REDCap data tibble
exists in a separate sheet.
}
\examples{
\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

supertbl <- read_redcap(redcap_uri, token)
write_redcap_xlsx(supertbl, file = "supertibble.xlsx")

write_redcap_xlsx(supertbl, labelled = TRUE, file = "supertibble.xlsx")
}

}
