% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_supertibble_xlsx}
\alias{write_supertibble_xlsx}
\title{Write Supertibbles to XLSX}
\usage{
write_supertibble_xlsx(
  supertbl,
  labelled = FALSE,
  labelled_sheets = FALSE,
  file,
  include_supertibble = TRUE,
  include_metadata = TRUE,
  tableStyle = "TableStyleLight10"
)
}
\arguments{
\item{supertbl}{a supertibble generated using \code{read_redcap()}}

\item{labelled}{Whether or not to include labelled outputs, default \code{FALSE}.
Requires use of \code{make_labelled()}.}

\item{labelled_sheets}{Whether or not to include labels in the XLSX sheets
instead of raw values. Default \code{FALSE}.}

\item{file}{A character string naming an xlsx file}

\item{include_supertibble}{Include a sheet capturing the supertibble output.
Default \code{TRUE}.}

\item{include_metadata}{Include a sheet capturing the combined output of the
supertibble \code{redcap_metadata}. Default \code{TRUE}.}

\item{tableStyle}{Any excel table style name or "none" (see "formatting"
vignette in \link[openxlsx]{writeDataTable}). Default "TableStyleLight10".}
}
\value{
A workbook object
}
\description{
Write supertibble outputs to XLSX file outputs, where each REDCap data tibble
exists in a separate sheet.
}
\examples{
\dontrun{
redcap_uri <- Sys.getenv("REDCAP_URI")
token <- Sys.getenv("REDCAP_TOKEN")

supertbl <- read_redcap(redcap_uri, token)
write_supertibble_xlsx(supertbl, file = "supertibble.xlsx")

write_supertibble_xlsx(supertbl, labelled = TRUE, file = "supertibble.xlsx")
}

}
